+++ Source/Mosa.Korlib/System/Decimal.cs
@@ -4,9 +4,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Diagnostics;
-using System.Globalization;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 namespace System
@@ -249,7 +246,7 @@ namespace System
 			this.lo = lo;
 			this.mid = mid;
 			this.hi = hi;
-			flags = ((int)scale) << 16;
+			flags = scale << 16;
 			if (isNegative)
 				flags |= SignMask;
 		}
